extract slf4-logger
