+++ subprojects/common/logger/src/main/kotlin/com/avito/logger/formatter/AppendClassNameFormatter.kt
@@ -1,0 +1,8 @@
+package com.avito.logger.formatter
+
+import com.avito.logger.LoggingFormatter
+
+class AppendClassNameFormatter(private val className: String) : LoggingFormatter {
+
+    override fun format(message: String): String = "[$className] $message"
+}
