+++ subprojects/common/slf4j-logger/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    id("kotlin")
+    `maven-publish`
+    id("com.jfrog.bintray")
+}
+
+dependencies {
+    api(project(":common:logger"))
+
+    implementation(Dependencies.slf4jApi)
+}
