+++ subprojects/common/slf4j-logger/src/main/kotlin/com/avito/logger/destination/Slf4jDestination.kt
@@ -5,7 +5,7 @@ import com.avito.logger.LoggingDestination
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
-internal class Slf4jDestination(private val tag: String) : LoggingDestination {
+class Slf4jDestination(private val tag: String) : LoggingDestination {
 
     @Transient
     private lateinit var _logger: Logger
