+++ subprojects/gradle/gradle-logger/build.gradle.kts
@@ -13,6 +13,7 @@ dependencies {
     implementation(project(":gradle:sentry-config"))
     implementation(project(":gradle:build-environment"))
     implementation(project(":common:elastic"))
+    implementation(project(":common:slf4j-logger"))
 
     testImplementation(Dependencies.Test.junitJupiterApi)
     testImplementation(Dependencies.Test.mockitoJUnitJupiter)
