+++ subprojects/settings.gradle.kts
@@ -78,6 +78,7 @@ include(":common:report-viewer")
 include(":common:report-viewer-test-fixtures")
 include(":common:elastic")
 include(":common:sentry")
+include(":common:slf4j-logger")
 include(":common:graphite")
 include(":common:graphite-test-fixtures")
 include(":common:statsd")
